src/test/java/ru/job4j/condition/RectangleAreaTest.java
package ru.job4j.condition;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.*;

class RectangleAreaTest {

    @Test
    void whenP7K2Square3Dot55() {
        float expected = 3.55f;
        int p = 8;
        double k = 2;
        double output = RectangleArea.square(p, k);
        assertThat(output).isEqualTo(expected, withPrecision(0.01));
    }

    @Test
    void whenP12K3Square6Dot75() {
        float expected = 6.75f;
        int p = 12;
        double k = 3;
        double output = RectangleArea.square(p, k);
        assertThat(output).isEqualTo(expected, withPrecision(0.01));
    }

    @Test
    void whenP3K1Square0Dot56() {
        float expected = 0.56f;
        int p = 3;
        double k = 1;
        double output = RectangleArea.square(p, k);
        assertThat(output).isEqualTo(expected, withPrecision(0.01));
    }
}